requirements.txt
requests==2.28.1
Pillow~=9.2.0
ruamel.yaml~=0.17.21
